Consolidate all classes from testing branch into single src/Main.java
Agent-Logs-Url: https://github.com/christienDuran/CompSci-3-Project/sessions/b0a2fa34-a29c-4221-9974-5e2e0ff0d8c2
