--- src/ProgressBar.java
@@ -1,2 +1,0 @@
-public class ProgressBar {
-}
