--- src/Reminder.java
@@ -1,16 +1,0 @@
-public class Reminder {
-
-
-    int reminderTime; // will be in format like 1130 or 1530
-    int snoozeLength; // in minutes most likely
-    String reminderDay; // maybe string for this? not sure yet
-    String type;  // event type
-
-    void sendAlert() {
-
-    }
-
-    void snoozeAlert() {
-
-    }
-}
