--- src/Settings.java
@@ -1,19 +1,0 @@
-public class Settings {
-    int theme;
-    String windowSize;
-    String timeFormat;
-    boolean sideBarOpen;
-
-
-    void changeTheme(int theme) {
-
-    }
-
-    void changeWindowSize(String windowSize) {}
-
-    void changeTimeFormat(String timeFormat) {}
-
-    void changeSideBarOpen(boolean sideBarOpen) {}
-
-
-}
