--- src/UserInputValidation.java
@@ -1,2 +1,0 @@
-public interface UserInputValidation {
-}
