--- src/WeatherService.java
@@ -1,18 +1,0 @@
-public class WeatherService {
-
-    double longitude;  // variables for location can be changed
-    double latitude;   // latitude and longitude is good for right now
-
-    double[] forecastData;  // we could have separate arrays for temp whether its sunny or cloudy chance of rain
-
-
-    void getWeather() {
-
-    }
-
-    void getForecastData() {
-
-    }
-
-
-}
